Add support for "create_as_copy" items
Those items will be copied instead of symlinked. Main use case is for
dotfiles which might be changed during the normal operation of the
application.

E.g. ~/.emacs shall be writeable so that it is still easy to quickly
install an emacs package and try it out. Same goes for customization.
Once things are in shape, they can be moved into the nix managed
configuration files. This should allow to have best of both.

Since the files are overwritten on each run, interim changes cannot pile
up over time ;-)